refactor(events): split events/tasks.py into a per-domain package (#575)
events/tasks.py sat at exactly 1000 lines — the hard CI file-length gate is
`-gt 1000`, so it passed by zero margin and the next edit would have forced an
unplanned split mid-PR. Split it pre-emptively into an events/tasks/ package
with domain-focused submodules (attendees, payments, invoicing, recurrence,
subscriptions, waitlist, exports, organization). Pure refactor, no behavior
change.

Celery task names are preserved exactly:
- Explicitly-named tasks keep their `name="events.X"`.
- The nine bare `@shared_task` tasks are pinned to `name="events.tasks.<func>"`
  (their previous default module-path name) so the registered name stays
  byte-identical across a rolling deploy. Verified at runtime: all 28 tasks
  register under their original names with zero submodule-path drift.

`events/tasks/__init__.py` re-exports every task, so all ~40 existing
`from events.tasks import <task>` import sites and task-function test patches
keep working untouched. The only test changes retarget two helper patches that
follow their task into a submodule: `events.tasks.send_email` →
`events.tasks.invoicing.send_email` and `events.tasks.httpx.post` →
`events.tasks.organization.httpx.post`.

Scoped to events/tasks.py only; the accounts/tasks.py half of #575 (which
additionally needs name-pinning for beat-scheduled bare tasks and the
email-wrapper consolidation) is left for a follow-up.

make check green (ruff, mypy --strict, migrations, i18n, file-length); 213
affected task/dispatcher tests pass.
